Added options for junit processing.
nojunit allows junit runs to be ignored (the test directory is still built).
junitlib specifies additional libs that are required through reflection.

This didn't acheive what I wanted: there is a circular dependency which is hidden
(through reflection).  The junit needs to float down separately from its strategem,
which is not currently supported.